Max transaction size limit is now accounted for when collating the signatures for a transaction. The first valid-collated transaction that is created is saved and used. Valid means that the key of the transaction has the required signatures, and that the transaction is within the size limitation.
When a transaction is signed, and prepared to be gathered and collated, the transaction being prepared was the post-signed transaction. This means that during the collate process, whichever actual .tx file that was used to collate all signatures would already have at least one signature already on it. Generally, this is not an issue, but if the amount of signatures pushed the total transaction size over the limit, then it could become an issue. Instead of actually signing the transaction along with creating the signature, the signature is now created and sent along side a non-signed transaction