Version bump
Collate helper's naming of files during collation wasn't quite right, especially when trying to handle the older and newer naming conventions.

ConvertToKeyCommand has been adjusted to not require a public key in order to create a pem/pub pair.

LargeBinaryFile was not properly signing the transactions, it was just creating txsig files. Now it is creating a sig and a tx file for each transaction created for the full file update. Also, the fileLink of the FileUpdate card in the home page was not working. Now it uses QuickLook in order to give some information of the content of the update.

Signed-off-by: John Bair <john.bair@swirldslabs.com>