When creating a FileUpdateTransaction, the contents were forced to be a zipped file. Now, it will take any file, and zip it if necessary. It will also remember if the file was originally zipped or not, to use the original file selected by the user when actually creating the transaction for signing and submitting.
When the real transaction is created, the contents are re-extracted from the original .lfu file in order to ensure it is the original contents. This is a temporary safety measure, as the original .lfu should not be alterable (OneDrive managed location), but the temporary files created to view the .lfu are.

Signed-off-by: John Bair <john.bair@swirldslabs.com>