Fixed some typos, including in a static String. Required fixing a few test files as well.
SubmitCommand was not properly ensuring that FileAppends were sent at the proper time (when the FileUpdate/Append before it reaches consensus and returns). Now it does. It will also work with multi-node submission as well.

Sign and Submit now for a FileUpdateTransaction being sent to multiple nodes was submitting each transaction one at a time, to each node. Now it will submit a transaction and its duplicates to all the nodes at the same time. Once any of the transactions gets consensus and returns, the next group of transactions will be submitted.

Signed-off-by: John Bair <john.bair@swirldslabs.com>